add constrain translation for subcircuit
