+++ align/compiler/match_graph.py
@@ -10,7 +10,7 @@ from ..schema.types import set_context
 import logging
 from ..schema import constraint
 from align.schema.graph import Graph
-
+from align.schema import ConstraintTranslator
 logger = logging.getLogger(__name__)
 
 
@@ -118,29 +118,6 @@ class Annotate:
         logger.debug(f"Subcircuits after creating primitive hiearchy {all_subckt}")
         return self.lib_names
 
-    def _check_const_length(self, const_list, const):
-        is_append = False
-        with set_context(const_list):
-            if hasattr(const, "instances") and len(const.instances) > 0:
-                is_append = True
-            elif (
-                isinstance(const, dict)
-                and "instances" in const
-                and len(const["instances"]) == 0
-            ):
-                # Modified constraint are initially of dict type
-                pass
-                # skipping const of zero length
-            elif not hasattr(const, "instances"):
-                is_append = True
-            else:
-                logger.debug(f"invalid constraint {const}")
-            if not is_append and const in const_list:
-                const_list.remove(const)
-            if is_append and const not in const_list:
-                logger.debug(f"constraint appended: {const}")
-                const_list.append(const)
-
     def _remove_group_const(self, subckt, rm_const):
         with set_context(subckt.constraints):
             start_count = len(subckt.constraints)
@@ -169,7 +146,7 @@ class Annotate:
             if isinstance(const, constraint.GroupBlocks)
         ]
         self._remove_group_const(subckt, gb_const)
-
+        tr = ConstraintTranslator(self.ckt_data)
         for const in gb_const:
             assert self.ckt_data.find(const.name.upper()) is None, "Already existing subckt with this name, please provide different name to const"
             const_inst = [i.upper() for i in const.instances]
@@ -221,14 +198,14 @@ class Annotate:
                 )
                 subckt.elements.append(X1)
             # Translate any constraints defined on the groupblock elements to subckt
-            self._top_to_bottom_translation(
+            tr._top_to_bottom_translation(
                 name, {inst: inst for inst in const_inst}, const.name
             )
             # Modify instance names in constraints after modifying groupblock
-            self._update_const(name, [const.name.upper(), *const_inst], inst_name)
+            tr._update_const(name, [const.name.upper(), *const_inst], inst_name)
         # Removing const with single instances.
         for c in list(self.ckt_data.find(name).constraints):
-            self._check_const_length(self.ckt_data.find(name).constraints, c)
+            tr._check_const_length(self.ckt_data.find(name).constraints, c)
         logger.debug(f"reduced constraints of design {name} {self.ckt_data.find(name).constraints}")
 
     def _group_cap_const(self, name):
@@ -243,7 +220,7 @@ class Annotate:
             logger.info(f"Existing GroupCaps constraint {gc_const} for subckt {name}")
         else:
             return
-
+        tr = ConstraintTranslator(self.ckt_data)
         for const in gc_const:
             for i in range(len(const.instances)):
                 const.instances[i] = const.instances[i].upper()
@@ -279,101 +256,7 @@ class Annotate:
                 )
                 subckt.elements.append(X1)
             # Modify instance names in constraints after modifying groupblock
-            self._update_const(
+            tr._update_const(
                 name, [const.name.upper(), *const_inst], const.name.upper()
             )
 
-    def _top_to_bottom_translation(self, top, match_dict, bottom):
-        """
-        Update instance names in the constraint in case they are reduced
-
-        Args:
-            top (str): name of subckt
-            match_dict (dict): node mapping
-        """
-
-        logger.debug(
-            f"transfering constraints from top subckt: {top} to bottom subckt: {bottom} "
-        )
-        assert self.ckt_data.find(bottom), f"Hierarchy not found, {bottom}"
-        const_list = self.ckt_data.find(top).constraints
-        sub_const = self.ckt_data.find(bottom).constraints
-        if not sub_const:
-            with set_context(sub_const):
-                for const in list(const_list):
-                    if any(
-                        isinstance(const, x)
-                        for x in [
-                            constraint.HorizontalDistance,
-                            constraint.VerticalDistance,
-                            constraint.BlockDistance,
-                            constraint.CompactPlacement,
-                        ]
-                    ):
-                        sub_const.append(const)
-                    elif hasattr(const, "instances"):
-                        # checking if sub hierarchy instances are in const defined
-                        sconst = {
-                            x: [
-                                match_dict[block]
-                                for block in const.instances
-                                if block in match_dict.keys()
-                            ]
-                            if x == "instances"
-                            else getattr(const, x)
-                            for x in const.__fields_set__
-                        }
-                        assert "constraint" in sconst
-                        # logger.debug(f"transferred constraint instances {match_dict} from {const} to {sconst}")
-                        self._check_const_length(
-                            self.ckt_data.find(bottom).constraints, sconst
-                        )
-                logger.debug(f"Transferred constraints to {bottom} {sub_const}")
-
-    def _update_const(self, name, remove_nodes, new_inst):
-        """
-        Update instance names in the constraint in case they are reduced by groupblock
-
-        Args:
-            name (str): name of subckt
-            G1 (graph): subckt graph
-            remove_nodes (list): nodes which are being removed
-        """
-
-        def _list_replace(lst, old_value, new_value):
-            for i, value in enumerate(lst):
-                if value == old_value:
-                    lst[i] = new_value
-
-        logger.debug(
-            f"update constraints at top hiearchy {name} :{remove_nodes} with {new_inst}"
-        )
-        const_list = self.ckt_data.find(name).constraints
-        for const in const_list:
-            if hasattr(const, "instances"):
-                # checking instances in the constraint and update names
-                if set(const.instances) & set(remove_nodes):
-                    replace = True
-                    for old_inst in remove_nodes:
-                        if replace:
-                            _list_replace(const.instances, old_inst, new_inst)
-                            replace = False
-                        elif old_inst in const.instances:
-                            const.instances.remove(old_inst)
-                    if len(const.instances) == 0:
-                        logger.debug(f"remove const belonging to new hierarchy {const}")
-                    # logger.debug(f"updated instances in the constraint:{const}")
-            elif hasattr(const, "pairs"):
-                for pair in const.pairs:
-                    if len(pair) == 2:
-                        if pair[0] in remove_nodes and pair[1] in remove_nodes:
-                            pair[0] = new_inst
-                            pair.pop()
-                        elif pair[0] in remove_nodes and pair[1] not in remove_nodes:
-                            pair[0] = new_inst
-                        elif pair[1] in remove_nodes and pair[0] not in remove_nodes:
-                            pair[1] = new_inst
-                    elif len(pair) == 1:
-                        if pair[0] in remove_nodes:
-                            pair[0] = new_inst
-        logger.debug(f"updated constraints of {name} {const_list}")
