+++ align/schema/__init__.py
@@ -5,3 +5,4 @@ from .instance import Instance
 from .library import Library
 from .parser import SpiceParser
 from .gen_dot import gen_dot_file
+from .translator import ConstraintTranslator
